+++ src/encoding/type_support/bstr.rs
@@ -75,7 +75,7 @@ impl DistinguishedValueDecoder<General> for bstr::BString {
 mod test {
     use super::{General, Vec};
     use crate::encoding::test::check_type_test;
-    check_type_test!(General, expedient, from Vec<u8>, into bstr::BString,
+    check_type_test!(General, relaxed, from Vec<u8>, into bstr::BString,
         WireType::LengthDelimited);
     check_type_test!(General, distinguished, from Vec<u8>, into bstr::BString,
         WireType::LengthDelimited);